Fixes mrope remove unecessary code
